feat: expand constant pool from 255 to 65535 (2-byte indices) (#76)
* feat: expand constant pool from 255 to 65535 (2-byte indices)

All constant-bearing opcodes now encode their pool index as a 2-byte
big-endian operand instead of 1 byte. ValueArray is backed by a
std::vector, the addConstant/getConstant API uses uint16_t, and every
emit/read site in the compiler, VM, and disassembler is updated
accordingly. Also fixes a latent GC bug: the memory_manager constant
marking loop used uint8_t as the loop variable, which would silently
wrap and miss constants at indices ≥ 256.


* style: apply clang-format to wide-constants changes


* refactor(bootstrap): remove runInterpreter() wrapper, no longer needed

The wrapper was added to spread constants across two chunks and stay
under the old 255-constant-per-chunk ceiling.  With the constant pool
expanded to 65535 the entire interpreter fits in one top-level chunk,
so remove the wrapper and replace the two early-exit `return` calls with
nested `if` guards (top-level `return` is a compile error in Lox++).


---------

+++ src/chunk.cpp
@@ -14,17 +14,17 @@ void Chunk::write(Op op, int line) { write(static_cast<Byte>(op), line); }
 
 void Chunk::patch(int offset, Byte byte) { (*this)[offset] = byte; }
 
-std::optional<uint8_t> Chunk::addConstant(Value value) {
-    // O(N) scan for an existing equal constant; avoids duplicate slots (N ≤
-    // 256). Strings are interned, so ObjHandle equality (index+type) suffices.
-    for (uint8_t i = 0; i < m_constants.size(); i++) {
+std::optional<uint16_t> Chunk::addConstant(Value value) {
+    // O(N) scan for an existing equal constant; avoids duplicate slots.
+    // Strings are interned, so ObjHandle equality (index+type) suffices.
+    for (uint16_t i = 0; i < m_constants.size(); i++) {
         if (m_constants.at(i) == value)
             return i;
     }
     if (m_constants.isFull())
         return std::nullopt;
     m_constants.write(value);
-    return static_cast<uint8_t>(m_constants.size() - 1);
+    return static_cast<uint16_t>(m_constants.size() - 1);
 }
 
 int Chunk::getLine(int offset) const {
@@ -38,4 +38,4 @@ int Chunk::getLine(int offset) const {
     return 0; // Error case
 }
 
-Value Chunk::getConstant(int idx) const { return m_constants.at(idx); }
+Value Chunk::getConstant(uint16_t idx) const { return m_constants.at(idx); }
